Apply autostart patch.
The autostart patch allows dwl to run an init script whenever dwl is
launched. This is convenient because running dwl with `-s` ruins the
standard output, which is needed for interacting with external programs:
`dwl -s /path/to/init/script > ~/.cache/dwltags`

However, the autostart patch allows us to run the program like this:
`dwl > ~/.cache/dwltags`

This means that the standard output can be redirected to the `dwltags`
file, where it can be read by script and fed into waybar.